fix(`btcdctrl`): handle & test for flaky edge-cases 1. killing a process then waiting for it can return an error depending on state, instead just exit the loop. 2. websockets hang indefinitely every 20 tests or so, use HTTP POST instead. 3. handle the RPC/P2P message being the last line in the log.
